Create additional module for args
