+++ db/db.json
@@ -1,1 +1,1 @@
-{"todos":[{"id":1,"task":"task1","description":"here is a description","deadline":"2003-04-03","progress":"undone"},{"id":2,"task":"task2","description":"-","deadline":"2003-04-03","progress":"undone"},{"id":3,"task":"task3","description":"fuck off","deadline":"2003-04-03","progress":"undone"}],"counter":3}
+{"todos":[{"id":1,"task":"task1","description":"here is a description","deadline":"2003-04-03","progress":"undone"},{"id":2,"task":"task2","description":"-","deadline":"2003-04-03","progress":"undone"},{"id":3,"task":"task3","description":"fuck off","deadline":"2003-04-03","progress":"undone"},{"id":4,"task":"do sth","description":"-","deadline":"2022-06-22","progress":"undone"}],"counter":4}
